requirements.txt
numpy==1.23.4
opencv-python==4.6.0.66
pdf2image==1.16.0
Pillow==9.2.0
pyinstaller==5.5
PyMuPDF==1.20.2
shutilwhich==1.1.0
